se refactoriza implementaciones
se refactoriza implementaciones
