+++ src/main/java/com/palomar/pokemoninfoservice/service/BitacoraService.java
@@ -1,21 +1,11 @@
 package com.palomar.pokemoninfoservice.service;
 
 import com.palomar.pokemoninfoservice.model.entity.Bitacora;
-import com.palomar.pokemoninfoservice.repository.BitacoraRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-@Service
-public class BitacoraService {
-    private final BitacoraRepository bitacoraRepository;
-
-    @Autowired
-    public BitacoraService(BitacoraRepository bitacoraRepository) {
-        this.bitacoraRepository = bitacoraRepository;
-    }
-
-    public Bitacora guardar(Bitacora bitacora) {
-        return bitacoraRepository.save(bitacora);
-    }
+/**
+ * Autor: josue.palomar
+ * Fecha de creación: 15/09/25
+ */
+public interface BitacoraService {
+    Bitacora guardar(Bitacora bitacora);
 }
-
