+++ src/main/java/com/palomar/pokemoninfoservice/service/PokemonApiService.java
@@ -1,21 +1,11 @@
 package com.palomar.pokemoninfoservice.service;
 
-import com.palomar.pokemoninfoservice.client.external.PokemonApiClient;
 import com.palomar.pokemoninfoservice.client.model.PokemonResponse;
-import com.palomar.pokemoninfoservice.model.entity.Bitacora;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-@Service
-public class PokemonApiService {
-    private final PokemonApiClient pokeApiClient;
-
-    @Autowired
-    public PokemonApiService(PokemonApiClient pokeApiClient) {
-        this.pokeApiClient = pokeApiClient;
-    }
-
-    public PokemonResponse getPokemonByName(String name) {
-        return pokeApiClient.getPokemonByName(name);
-    }
+/**
+ * Autor: josue.palomar
+ * Fecha de creación: 15/09/25
+ */
+public interface PokemonApiService {
+    public PokemonResponse getPokemonByName(String name);
 }
